+++ src/agent_ddpg.py
@@ -10,22 +10,22 @@ from src.utils import Memory
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
   
 class agent_ddpg:
-    def __init__(self, env, hidden_size=[400, 300], actor_learning_rate=1e-4, critic_learning_rate=1e-3, gamma=0.99, tau=1e-3, max_memory_size=50000, discrete = [0, 10, 0], norm = 'none'):
+    def __init__(self, env, hidden_size=[400, 300], actor_learning_rate=1e-4, critic_learning_rate=1e-3, gamma=0.99, tau=1e-3, max_memory_size=50000, norm = 'none'):
         
         #BiddingMarket_energy_Environment Params
-        self.discrete = discrete
         self.num_states = env.observation_space.shape[0]
         self.num_actions = env.action_space.shape[0]
-
+        
+        # DDPG specific Params
         self.gamma = gamma
         self.tau = tau
         self.norm = norm
         self.hidden_size= hidden_size
-        self.output_size = 1 #only for critic; before, output size was equal to action size (self.num_actions)
+        self.output_size = 1 #only for critic
         
         # Networks
-        self.actor = Actor(self.num_states, self.hidden_size, self.num_actions, discrete = self.discrete, norm = self.norm).to(device)
-        self.actor_target = Actor(self.num_states, self.hidden_size, self.num_actions, discrete = self.discrete , norm = self.norm).to(device)
+        self.actor = Actor(self.num_states, self.hidden_size, self.num_actions, norm = self.norm).to(device)
+        self.actor_target = Actor(self.num_states, self.hidden_size, self.num_actions, norm = self.norm).to(device)
         self.critic = Critic(self.num_states, self.hidden_size, self.output_size, self.num_actions, norm = self.norm).to(device)
         self.critic_target = Critic(self.num_states, self.hidden_size, self.output_size, self.num_actions, norm = self.norm).to(device)
 
@@ -43,51 +43,36 @@ class agent_ddpg:
     
     def get_action(self, state):
         state = Variable(torch.from_numpy(state).float().unsqueeze(0).to(device))
-        self.actor.eval() #new
+        self.actor.eval()
         action = self.actor.forward(state)
-        self.actor.train() #new
+        self.actor.train() 
         #action = action.detach().numpy()[0,:]  
         action = action.detach().cpu().numpy()[0,:]  
         return action
     
     def update(self, batch_size):
+        
         states, actions, rewards, next_states, _ = self.memory.sample(batch_size)
-        
-        # depends on update Versions of pytorch and conda
-        #states = torch.cuda.FloatTensor(states, device=device)
-        #actions = torch.cuda.FloatTensor(actions, device=device)
-        #rewards = torch.cuda.FloatTensor(rewards, device=device)
-        #next_states = torch.cuda.FloatTensor(next_states, device=device)
-        
         states = torch.FloatTensor(states, device=device)
         actions = torch.FloatTensor(actions, device=device)
         rewards = torch.FloatTensor(rewards, device=device)
         next_states = torch.FloatTensor(next_states, device=device)
-        
-        # depends on update Versions of pytorch and conda
-        #states = torch.from_numpy(states).float().to(device)
-        #actions = torch.from_numpy(actions).float().to(device)
-        #rewards = torch.from_numpy(rewards).float().to(device)
-        #next_states = torch.from_numpy(next_states).float().to(device)
-        
         
         # Critic loss       
         Qvals = self.critic.forward(states, actions)
         next_actions = self.actor_target.forward(next_states)
         next_Q = self.critic_target.forward(next_states, next_actions.detach())
         Qprime = rewards + self.gamma * next_Q
-        #critic_loss = self.critic_criterion(Qvals, Qprime) 
-
-        # Actor loss
-        #policy_loss = -self.critic.forward(states, self.actor.forward(states)).mean()
         
         # update networks
         self.critic_optimizer.zero_grad()
+        # Critic loss
         critic_loss = self.critic_criterion(Qvals, Qprime)
         critic_loss.backward() 
         self.critic_optimizer.step()
         
         self.actor_optimizer.zero_grad()
+        # Actor loss
         policy_loss = -self.critic.forward(states, self.actor.forward(states)).mean()
         policy_loss.backward()
         self.actor_optimizer.step()