+++ src/utils.py
@@ -1,19 +1,16 @@
 import numpy as np
 import random
+import matplotlib.pyplot as plt
 from collections import deque
 
 
-# Ornstein-Ulhenbeck Process, Taken from #https://github.com/vitchyr/rlkit/blob/master/rlkit/exploration_strategies/ou_strategy.py
-
-
 class UniformNoise(object):
-    def __init__(self, action_space, price_cap, initial_exploration = 0.99, final_exploration = 0.05, decay_rate = 0.999):
+    def __init__(self, action_space, initial_exploration = 0.99, final_exploration = 0.05, decay_rate = 0.999):
         
         self.action_dim      = action_space.shape[0] # Requires Space with (10,) shape!
         self.low             = action_space.low
         self.high            = action_space.high
         self.distance        = abs(self.low - self.high)
-        self.price_cap       = price_cap
         
         self.initial_exploration = initial_exploration
         self.final_exploration   = final_exploration
@@ -32,7 +29,7 @@ class UniformNoise(object):
         explore_yes = np.random.binomial(1,exploration_probabilty)
          
         # Unnormalized Uniform Numbers
-        noise_list = np.random.uniform(self.low, self.price_cap ,size=self.action_dim) #used self.low/10 before
+        noise_list = np.random.uniform(self.low, self.high ,size=self.action_dim) #used self.low/10 before
         
         #Renormalize
         #sum_noise = noise_list.sum()
@@ -40,8 +37,7 @@ class UniformNoise(object):
         
         return noisy_action 
     
-    
-
+# Ornstein-Ulhenbeck Process, Taken from #https://github.com/vitchyr/rlkit/blob/master/rlkit/exploration_strategies/ou_strategy.py
 class OUNoise(object):
     def __init__(self, action_space, mu=0.0, theta=0.15, max_sigma=0.3, min_sigma=0.3, decay_period=100000):
         self.mu           = mu
@@ -148,3 +144,74 @@ class GaussianNoise(object):
         noisy_action = np.clip(action + noise_list, self.low, self.high)
 
         return noisy_action 
+
+
+    
+def plot_run_outcome(data, number_of_agents, bid_limit, NE, episodes, run, curves = 'both', 
+                       title = 'Bidding Game',rescale=[1,1,1], moving_window = 9):
+    '''
+    Plots actions or rewards or both (curves ='actons','rewards'or'both')
+    Reads out nessecary data from dictionary (structured like in 'main')
+    possible to display bid limit and Nash Equilibrium (NE) threshold (NE can also be 'none')
+    actions, rewards and bid_limit can be recaled for reprenstatiom (rescale[param for actions, param, for rewards, param for bid_limit])
+    also takes the moving medain of a predefined window (default = 10) for smoothing outputted lines
+    '''
+    
+    med_actions, med_rewards = moving_median_rewards_actions(data,run,episodes, moving_window)
+    
+    # rescale data
+    med_actions = med_actions*rescale[0]
+    med_rewards = med_rewards*rescale[1]
+    bid_limit = bid_limit*rescale[2] 
+    
+    plt.plot([bid_limit]*episodes, color='grey',label = 'Bid Limit' , lw =1)
+    
+    if NE != 'none':
+        plt.plot([NE]*episodes, color='C0', label = 'Nash Equilibrium', lw =1)
+    
+    for i in range(number_of_agents):
+        if curves == 'actions':
+            plt.plot(med_actions[1:,i], 'C{}'.format(i+1), label = 'Bids Agent{}'.format(i), lw =1, linestyle = '--') # displaying actions
+            plt.ylabel('Action')
+        elif curves == 'rewards':
+            plt.plot(med_rewards[1:,i], 'C{}'.format(i+1), label = 'Rewards Agent1', lw =1) # displaying rewards
+            plt.ylabel('Reward')
+        else:
+            plt.plot(med_actions[1:,i], 'C{}'.format(i+1), label = 'Bids Agent{}'.format(i), lw =1, linestyle = '--')
+            plt.plot(med_rewards[1:,i], 'C{}'.format(i+1), label = 'Rewards Agent1', lw =1) # displaying rewards
+            plt.ylabel('Reward/Action')
+
+
+    
+    plt.xlabel('Episode')
+    plt.legend(loc=4, prop={'size': 7})
+    plt.title('{}'.format(title))
+    plt.plot()
+    
+
+
+def moving_median_rewards_actions(data,run, episodes=15000, n=9): 
+    '''
+    reads actions and rewards for all episodes in a specific run from dictionary 
+    and further calculates the moving median for a specified period "n"
+    '''
+    # get data from dictionary
+    actions =[data[run][i]['actions'] for i in range(episodes)]
+    rewards =[data[run][i]['rewards'] for i in range(episodes)]
+    actions= np.squeeze(np.asarray(actions))
+    rewards= np.squeeze(np.asarray(rewards))
+    
+    med_rewards = []
+    med_actions = []
+    #moving median of n_th step (takes median from n rows and outputs an array of the same length as the input)
+    for i in range(episodes):
+        temp_actions =np.median(actions[i:n], axis=0)
+        temp_rewards =np.median(rewards[i:n], axis=0)
+        med_actions.append(temp_actions)
+        med_rewards.append(temp_rewards)
+        n += 1
+
+    recompiled_actions = np.asarray(med_actions)
+    recomplied_rewards = np.asarray(med_rewards)
+    
+    return recompiled_actions, recomplied_rewards